Support `--foo=bar` syntax (#98)
Fixes https://github.com/com-lihaoyi/mainargs/issues/97

Relatively straightforward change. We do `head.split("=", 2) match` to
see if it contains any `=`, and if so treat the portion of the string
after the first `=` as the value. Added some unit test to cover both
success cases and failure cases: use with short names or with flags is
unsupported.