Narrow the return type of a few implicits
This came up when upgrading our code base from an older vesion of mainargs, where we relied on chaining TokenReaders for Map and Seq. The implicit conversions require `TokensReader.Simple` as argument, but return the superclass `TokensReader`, which makes chaining impossible. Since the implementation is actually using the narrow type, we just expose it in the implicit def signature to make chaining work again.